soporte para notas sostenidas en engine nuevo

+++ Assets/Cupi/Scripts/Gameplay/tilesMode/TiposNotas/NotaTileBaseLogic.cs
@@ -14,7 +14,7 @@ public class NotaTileBaseLogic : MonoBehaviour
     public Vector2 DireccionMovimiento;
     public Vector2 finalPos;
     public Transform origin;
-    public float progress;
+    public double progress;
     public bool lockProgress;
     public float offsetRendering;
     public int Myindex;
@@ -49,13 +49,13 @@ public class NotaTileBaseLogic : MonoBehaviour
     {
         if (!initialized) return;
 
-        progress = 1 - InverseLerpUnclamped(0.0f, data.timeToArrive + offsetRendering, (float)timeController.AdditiveTime);
+        progress = 1 - InverseLerpUnclamped(0.0f, data.timeToArrive + offsetRendering, timeController.AdditiveTime);
 
         if (lockProgress) progress = Mathf.Max(0);
 
-        float distancia = (progress * (data.timeToArrive + offsetRendering) * data.localSpeed * tilesModeMaster.NotaTileSpeed);
+        double distancia = (progress * (data.timeToArrive + offsetRendering) * data.localSpeed * tilesModeMaster.NotaTileSpeed);
 
-        finalPos = data.offsetPositionToGo + (DireccionMovimiento * distancia);
+        finalPos = data.offsetPositionToGo + (DireccionMovimiento * (float)distancia);
 
         note.localPosition = finalPos;
 
@@ -76,7 +76,7 @@ public class NotaTileBaseLogic : MonoBehaviour
         GoToPool();
     }
 
-    public float InverseLerpUnclamped(float a, float b, float valor)
+    public double InverseLerpUnclamped(double a, double b, double valor)
     {
         if (b != a) return (valor - a) / (b - a);
